♻️ | AB-8 | REFACTORING | Extract categories in new folder constants
